refactor: renamed ssh-server to sftp-server
